Fix generic compiler warnings

+++ V-Engine/Clock.cpp
@@ -13,10 +13,10 @@ Time Clock::GetDeltatime()
 	const auto old = m_last;
 	m_last = steady_clock::now();
 	const duration<double> delta = m_last - old;
-	return Time(delta.count());
+	return Time( static_cast<float>( delta.count() ) );
 }
 
 Time Clock::PeekDeltatime() const
 {
-	return duration<double>(steady_clock::now() - m_last).count();
+	return duration<float>(steady_clock::now() - m_last).count();
 }
